+++ tests/testthat/setup.R
@@ -916,7 +916,7 @@ options(projpred.warn_cvrefbuilder_NULL = FALSE)
 # Suppress warnings thrown while fitting the submodels:
 options(projpred.warn_proj_drawwise = FALSE)
 # Don't use the convergence checker:
-options(projpred.check_conv = FALSE)
+options(projpred.check_convergence = FALSE)
 # Set default number of significant digits to be printed:
 options(projpred.digits = getOption("digits"))